wifi-scripts: Add caldata support for al02-c11
Add caldata support for al02-c11

+++ wifi-scripts/files/read_caldata_to_fs.sh
@@ -273,17 +273,19 @@ do_load_ipq4019_board_bin()
                     cp ${apdk}/IPQ6018/caldata.bin /lib/firmware/IPQ6018/caldata.bin
             ;;
             ap-al02-c4*|ap-al02-c7*|ap-al02-c8*)
+                    [ -f /lib/firmware/qcn9224/caldata_2.bin ] && exit 1;
                     WKK_FILESIZE=184320
                     mkdir -p ${apdk}/qcn9224
 
                     dd if=${mtdblock} of=${apdk}/qcn9224/caldata_2.bin bs=1 count=$WKK_FILESIZE skip=4096
                     dd if=${mtdblock} of=${apdk}/qcn9224/caldata_3.bin bs=1 count=$WKK_FILESIZE skip=208896
                     dd if=${mtdblock} of=${apdk}/qcn9224/caldata_4.bin bs=1 count=$WKK_FILESIZE skip=413696
-                    ln -s ${apdk}/qcn9224/caldata_2.bin /lib/firmware/qcn9224/caldata_2.bin
-                    ln -s ${apdk}/qcn9224/caldata_3.bin /lib/firmware/qcn9224/caldata_3.bin
-                    ln -s ${apdk}/qcn9224/caldata_4.bin /lib/firmware/qcn9224/caldata_4.bin
+                    cp ${apdk}/qcn9224/caldata_2.bin /lib/firmware/qcn9224/caldata_2.bin
+                    cp ${apdk}/qcn9224/caldata_3.bin /lib/firmware/qcn9224/caldata_3.bin
+                    cp ${apdk}/qcn9224/caldata_4.bin /lib/firmware/qcn9224/caldata_4.bin
             ;;
             ap-al02-c6*)
+                    [ -f /lib/firmware/IPQ9574/caldata.bin ] && exit 1;
                     AL_BD_FILENAME=/lib/firmware/IPQ9574/bdwlan.bin
                     mkdir -p ${apdk}/IPQ9574
                     if [ -f "$AL_BD_FILENAME" ]; then
@@ -293,23 +295,42 @@ do_load_ipq4019_board_bin()
                     fi
                     dd if=${mtdblock} of=${apdk}/IPQ9574/caldata.bin bs=1 count=$FILESIZE skip=4096
                     [ -L /lib/firmware/IPQ9574/caldata.bin ] || \
-                    ln -s ${apdk}/IPQ9574/caldata.bin /lib/firmware/IPQ9574/caldata.bin
+                    cp ${apdk}/IPQ9574/caldata.bin /lib/firmware/IPQ9574/caldata.bin
 
                     WKK_FILESIZE=184320
                     mkdir -p ${apdk}/qcn9224
                     dd if=${mtdblock} of=${apdk}/qcn9224/caldata_3.bin bs=1 count=$WKK_FILESIZE skip=208896
                     dd if=${mtdblock} of=${apdk}/qcn9224/caldata_4.bin bs=1 count=$WKK_FILESIZE skip=413696
-                    ln -s ${apdk}/qcn9224/caldata_3.bin /lib/firmware/qcn9224/caldata_3.bin
-                    ln -s ${apdk}/qcn9224/caldata_4.bin /lib/firmware/qcn9224/caldata_4.bin
+                    cp ${apdk}/qcn9224/caldata_3.bin /lib/firmware/qcn9224/caldata_3.bin
+                    cp ${apdk}/qcn9224/caldata_4.bin /lib/firmware/qcn9224/caldata_4.bin
             ;;
             ap-al02-c9*)
+                    [ -f /lib/firmware/qcn9224/caldata_1.bin ] && exit 1;
                     WKK_FILESIZE=184320
                     mkdir -p ${apdk}/qcn9224
 
                     dd if=${mtdblock} of=${apdk}/qcn9224/caldata_1.bin bs=1 count=$WKK_FILESIZE skip=4096
                     dd if=${mtdblock} of=${apdk}/qcn9224/caldata_3.bin bs=1 count=$WKK_FILESIZE skip=208896
-                    ln -s ${apdk}/qcn9224/caldata_1.bin /lib/firmware/qcn9224/caldata_1.bin
-                    ln -s ${apdk}/qcn9224/caldata_3.bin /lib/firmware/qcn9224/caldata_3.bin
+                    cp ${apdk}/qcn9224/caldata_1.bin /lib/firmware/qcn9224/caldata_1.bin
+                    cp ${apdk}/qcn9224/caldata_3.bin /lib/firmware/qcn9224/caldata_3.bin
+            ;;
+            ap-al02-c11*)
+                    [ -f /lib/firmware/IPQ9574/caldata.bin ] && exit 1;
+                    AL_BD_FILENAME=/lib/firmware/IPQ9574/bdwlan.bin
+                    mkdir -p ${apdk}/IPQ9574
+                    if [ -f "$AL_BD_FILENAME" ]; then
+                        FILESIZE=$(stat -Lc%s "$AL_BD_FILENAME")
+                    else
+                        FILESIZE=131072
+                    fi
+                    dd if=${mtdblock} of=${apdk}/IPQ9574/caldata.bin bs=1 count=$FILESIZE skip=4096
+                    [ -L /lib/firmware/IPQ9574/caldata.bin ] || \
+                    cp ${apdk}/IPQ9574/caldata.bin /lib/firmware/IPQ9574/caldata.bin
+
+                    WKK_FILESIZE=184320
+                    mkdir -p ${apdk}/qcn9224
+                    dd if=${mtdblock} of=${apdk}/qcn9224/caldata_3.bin bs=1 count=$WKK_FILESIZE skip=208896
+                    cp ${apdk}/qcn9224/caldata_3.bin /lib/firmware/qcn9224/caldata_3.bin
             ;;
             ap-al02*)
                     [ -f /lib/firmware/IPQ9574/caldata.bin ] && exit 1;
